Updated demo. Also changed ClientID to be of type long in attempt to fix the clientId sync issue.

+++ src/Ycs/Structs/AbstractStruct.cs
@@ -26,7 +26,7 @@ namespace Ycs
         internal abstract bool MergeWith(AbstractStruct right);
         internal abstract void Delete(Transaction transaction);
         internal abstract void Integrate(Transaction transaction, int offset);
-        internal abstract int? GetMissing(Transaction transaction, StructStore store);
+        internal abstract long? GetMissing(Transaction transaction, StructStore store);
         internal abstract void Write(IUpdateEncoder encoder, int offset);
     }
 }
