+++ src/Ycs/Utils/Snapshot.cs
@@ -14,9 +14,9 @@ namespace Ycs
     public sealed class Snapshot : IEquatable<Snapshot>
     {
         internal readonly DeleteSet DeleteSet;
-        internal readonly IDictionary<int, int> StateVector;
+        internal readonly IDictionary<long, int> StateVector;
 
-        internal Snapshot(DeleteSet ds, IDictionary<int, int> stateMap)
+        internal Snapshot(DeleteSet ds, IDictionary<long, int> stateMap)
         {
             DeleteSet = ds;
             StateVector = stateMap;
@@ -40,7 +40,7 @@ namespace Ycs
                     // Splitting the structs before writing them to the encoder.
                     foreach (var kvp in StateVector)
                     {
-                        int client = kvp.Key;
+                        long client = kvp.Key;
                         int clock = kvp.Value;
 
                         if (clock == 0)
