+++ src/Ycs/Utils/YDoc.cs
@@ -330,7 +330,7 @@ namespace Ycs
             using (var encoder = new UpdateEncoderV2())
             {
                 var targetStateVector = encodedTargetStateVector == null
-                    ? new Dictionary<int, int>()
+                    ? new Dictionary<long, int>()
                     : EncodingUtils.DecodeStateVector(new MemoryStream(encodedTargetStateVector, writable: false));
                 WriteStateAsUpdate(encoder, targetStateVector);
                 return encoder.ToArray();
@@ -350,7 +350,7 @@ namespace Ycs
         /// Write all the document as a single update message. If you specify the satte of the remote client, it will only
         /// write the operations that are missing.
         /// </summary>
-        internal void WriteStateAsUpdate(IUpdateEncoder encoder, IDictionary<int, int> targetStateVector)
+        internal void WriteStateAsUpdate(IUpdateEncoder encoder, IDictionary<long, int> targetStateVector)
         {
             EncodingUtils.WriteClientsStructs(encoder, Store, targetStateVector);
             new DeleteSet(Store).Write(encoder);
